move around pipeline types a bit
