+++ fiftyone/operators/_types/__init__.py
@@ -1,0 +1,7 @@
+"""
+Private subpackage for FiftyOne operator types.
+
+| Copyright 2017-2025, Voxel51, Inc.
+| `voxel51.com <https://voxel51.com/>`_
+|
+"""
