--- fiftyone/operators/pipeline.py
@@ -1,65 +1,0 @@
-"""
-FiftyOne pipeline operator.
-
-| Copyright 2017-2025, Voxel51, Inc.
-| `voxel51.com <https://voxel51.com/>`_
-|
-"""
-import dataclasses
-from typing import Any, Mapping, Optional
-
-from .operator import Operator
-
-
-@dataclasses.dataclass
-class PipelineStage:
-    """A stage in a FiftyOne pipeline.
-
-    Args:
-        operator_uri: the URI of the operator to use for the stage
-        name: the name of the stage
-        num_distributed_tasks: the number of distributed tasks to use
-            for the stage, optional
-        params: optional parameters to pass to the operator, overwriting
-            any existing parameters
-    """
-
-    # Required
-    operator_uri: str
-
-    # Optional
-    name: Optional[str] = None
-    num_distributed_tasks: Optional[int] = None
-    params: Optional[Mapping[str, Any]] = None
-
-    def __post_init__(self):
-        if not self.operator_uri:
-            raise ValueError("operator_uri must be a non-empty string")
-
-        if (
-            self.num_distributed_tasks is not None
-            and self.num_distributed_tasks < 1
-        ):
-            raise ValueError("num_distributed_tasks must be >= 1")
-
-
-class PipelineOperator(Operator):
-    def resolve_pipeline(self, ctx):
-        """Returns the resolved pipeline of the operator.
-
-        Subclasses can implement this method to define the pipeline of the
-        operator.
-
-        Args:
-            ctx: the :class:`fiftyone.operators.executor.ExecutionContext`
-
-        Returns:
-            a list of :class:`fiftyone.operators.PipelineStage`,
-        """
-        raise NotImplementedError("subclass must implement resolve_pipeline")
-
-    def execute(self, ctx):
-        """Not used for pipeline operators; pipelines are executed via stages"""
-        raise NotImplementedError(
-            "execute() not implemented for PipelineOperators"
-        )
